+++ modules/special_functions/src/include/InterpolationND.hpp
@@ -16,19 +16,7 @@
 namespace Nelson {
 //=============================================================================
 NLSSPECIAL_FUNCTIONS_IMPEXP ArrayOf
-LinearInterpolation1D(const ArrayOf& X, const ArrayOf& V, const ArrayOf& XQ);
-//=============================================================================
-NLSSPECIAL_FUNCTIONS_IMPEXP ArrayOf
-LinearInterpolation1D(const ArrayOf& V, const ArrayOf& XQ);
-//=============================================================================
-NLSSPECIAL_FUNCTIONS_IMPEXP ArrayOf
-LinearInterpolation1D(const ArrayOf& X, const ArrayOf& V, const ArrayOf& XQ,
-    const std::wstring& methodName, const std::wstring& extrapolationMode,
-    const ArrayOf* extrapolationValue = nullptr);
-//=============================================================================
-NLSSPECIAL_FUNCTIONS_IMPEXP ArrayOf
-LinearInterpolation1D(const ArrayOf& V, const ArrayOf& XQ, const std::wstring& methodName,
-    const std::wstring& extrapolationMode, const ArrayOf* extrapolationValue = nullptr);
+InterpolationND(const ArrayOfVector& argIn);
 //=============================================================================
 }
 //=============================================================================